src/pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "torch"]
build-backend = "setuptools.build_meta"

[project]
name = "clairai"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x",
    "numpy",
]